pyproject.toml
[project]
name = "sf10x"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langchain-community",
    "loguru",
    "jq",
    "langchain-google-genai",
    "langchain-core",
    "faiss-cpu",
    "langchain-openai",
    "python-dotenv",
    "langchain-anthropic",
    "fastapi==0.115.6",
    "pydantic~=2.10.4",
    "uvicorn==0.34.0",
    "streamlit",
    "chonkie[model2vec]",
    "tiktoken",
    "sqlalchemy~=2.0.0",
    "asyncpg",
    "psycopg2-binary",
    "alembic",
    "pgvector",
    "greenlet",
    "requests",
    "beautifulsoup4",
    "lxml",
    "supabase",
    "aiohttp",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]